backend/requirements.txt
Django==3.2.15
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
Pillow==9.2.0
PyJWT==2.4.0
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
